fix test runner to fail with error count (#51)
I noticed that if TestEZ encounters a require error, then it does 
not increase the failureCount from the test run results (probably
because it does not happen directly in a test) so the CI would still
pass. With that fix, the test run will exit with non-zero for any 
errors that TestEZ catches.